// src/generated/prisma/enums.ts
export const TxnType = {
  DEPOSIT: 'DEPOSIT',
  EARNING: 'EARNING',
  PAYOUT: 'PAYOUT',
  PLATFORM_FEE: 'PLATFORM_FEE',
  REFUND: 'REFUND',
  MEMBERSHIP_FEE: 'MEMBERSHIP_FEE',
  REFERRAL_BONUS: 'REFERRAL_BONUS',
  ADMIN_CREDIT: 'ADMIN_CREDIT',
  LOAN: 'LOAN',
  MILESTONE_BONUS: 'MILESTONE_BONUS',
  GTMS_PASS_FEE: 'GTMS_PASS_FEE',
  WELCOME_BONUS: 'WELCOME_BONUS'
} as const

export type TxnType = (typeof TxnType)[keyof typeof TxnType]